+++ server/constants_app.go
@@ -2,5 +2,5 @@ package server
 
 const (
 	Name    = "upda"
-	Version = "1.0.1"
+	Version = "1.0.2"
 )